backend/image-transform/requirements.txt
# Image Transformation Lambda Dependencies
# These will be packaged in a Lambda Layer for reuse

# Core
boto3>=1.28.0

# Image processing
Pillow>=10.0.0
rawpy>=0.18.0
pillow-heif>=0.13.0
numpy>=1.24.0

